tests/requirements.txt
requests>=2.28
Pillow>=9.0
pytest>=7.0
